fix(no-jira): Fixing flaky Dashboard test and moving it  into postMH stage
